requirements.txt
flask==3.1.0
discord.py==2.4.0
redis==5.2.1
python-dotenv==1.0.1
